refactor: extract topologyAssembler + topologyTypes from network/service.ts
Closes #973 partially.

Pulls two structural seams out of the 2107-LOC monolith:

- topologyTypes.ts — shared port helpers and the duck-typed shapes
  used by both the global subgraph builder and the per-node assembly:
  resolvePortNumber, KubePodSpec, PortLike, FritzPortMapping.
- topologyAssembler.ts — buildGlobalInfrastructure(), the
  always-visible Internet + Gateway + external-link triangle that
  every per-node subgraph hangs off. Was the private
  getGlobalInfrastructure method on NetworkService.

NetworkService.getGraph now calls buildGlobalInfrastructure()
directly; no wrapper needed. The flowSampler driver move and the
domainReachabilityCache extraction stay open for follow-ups — they
need the per-node subgraph split (still 1600+ LOC of getNodeGraph)
to come first.

Sizes:
- service.ts: 2107 to 1946 LOC
- topologyAssembler.ts: 155 LOC (new)
- topologyTypes.ts: 75 LOC (new)

Zero behaviour change verified by the full backend test suite incl.
graph.test.ts (9 tests pinning the global infrastructure shape +
external-link edge resolution).
